+++ NickolaiLychagin/rss_reader/rss_reader_colors.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3
+
+"""Colors for colorization"""
+
+COLORS = {"red": "\033[91m",
+          "green": "\033[92m",
+          "yellow": "\033[93m",
+          "blue": "\033[94m",
+          "magenta": "\033[95m",
+          "cyan": "\033[96m",
+          "reset": "\033[0m"}